fix(quality): resolve all review findings — blockers, security, arch
- agent_refiner_stages: catch (json.JSONDecodeError, TypeError) — json.loads raises both
- cli_nats: type op param as Literal in _build_control_payload signature, drop cast
- discord_outbound: add StageChannel to _PartialMessageable; assert → explicit if guard
- cli_pool_types: real defaults (True/"") instead of Ellipsis; remove __all__; add TYPE_CHECKING note
- cli_pool_streaming: remove unnecessary _pool_id alias in _reset closure
- bootstrap_stores: validate idx_sql against allowlist regex before DDL replay
- auth.conf: remove $JS.API.STREAM.CREATE/UPDATE.LYRA_AUDIT from hub publish
- check-acl-authconf-drift: rename TMPDIR_LOCAL→NATS_TMPDIR; strip inline comments; safe diff capture
- check-inbox-prefix: fix grep -z portability; replace docstring heuristic with Python AST check
- trace: TraceLogRecord → Protocol (never instantiated; subclass was semantically wrong)
- cli_protocol_types: extract shared types to break circular import; cli_protocol is now a clean facade (ADR-060)
- discord_normalize/telegram_normalize: restore is_admin comment to trailing position
- discord_audio_outbound: document discord.http.Route as internal API
- cli_setup: log.debug → log.warning for plugin load failure
- middleware_pool: drop getattr fallback; document _configured idempotency invariant
- pool: annotate surviving BLE001 catches with reason tags
- auth.conf: document llm-worker/monitor no-inbox intent; document _inbox lowercase variants
- tools: add file/folder exemptions for bootstrap_stores (+5 lines security) and cli/ (+1 file ADR-060)
