+++ deploy/nats/auth.conf
@@ -13,7 +13,7 @@ authorization {
       nkey: "UDUMMYHUB"
       # hub
       permissions: {
-        publish:   { allow: ["lyra.outbound.telegram.>","lyra.outbound.discord.>","lyra.voice.tts.request.>","lyra.voice.stt.request.>","lyra.llm.request","lyra.image.generate.request","lyra.clipool.cmd","lyra.clipool.control","$JS.API.STREAM.CREATE.LYRA_AUDIT","$JS.API.STREAM.UPDATE.LYRA_AUDIT","lyra.audit.>"] }
+        publish:   { allow: ["lyra.outbound.telegram.>","lyra.outbound.discord.>","lyra.voice.tts.request.>","lyra.voice.stt.request.>","lyra.llm.request","lyra.image.generate.request","lyra.clipool.cmd","lyra.clipool.control","lyra.audit.>"] }
         subscribe: { allow: ["lyra.inbound.telegram.>","lyra.inbound.discord.>","lyra.voice.tts.heartbeat","lyra.voice.stt.heartbeat","lyra.llm.health.*","lyra.system.ready","_INBOX.hub.>","lyra.image.heartbeat","lyra.clipool.heartbeat"] }
         allow_responses: true
       }
@@ -41,7 +41,7 @@ authorization {
       # tts-adapter
       permissions: {
         publish:   { allow: ["lyra.voice.tts.heartbeat","lyra.system.ready"] }
-        subscribe: { allow: ["lyra.voice.tts.request","lyra.voice.tts.request.>","_INBOX.tts-adapter.>","_inbox.tts-adapter.>"] }
+        subscribe: { allow: ["lyra.voice.tts.request","lyra.voice.tts.request.>","_INBOX.tts-adapter.>","_inbox.tts-adapter.>"] }  # lowercase variant for nats.py <2.2 compatibility
         allow_responses: true
       }
     }
@@ -50,7 +50,7 @@ authorization {
       # stt-adapter
       permissions: {
         publish:   { allow: ["lyra.voice.stt.heartbeat","lyra.system.ready"] }
-        subscribe: { allow: ["lyra.voice.stt.request","lyra.voice.stt.request.>","_INBOX.stt-adapter.>","_inbox.stt-adapter.>"] }
+        subscribe: { allow: ["lyra.voice.stt.request","lyra.voice.stt.request.>","_INBOX.stt-adapter.>","_inbox.stt-adapter.>"] }  # lowercase variant for nats.py <2.2 compatibility
         allow_responses: true
       }
     }
@@ -59,7 +59,7 @@ authorization {
       # voice-tts
       permissions: {
         publish:   { allow: ["lyra.voice.tts.heartbeat","lyra.system.ready"] }
-        subscribe: { allow: ["lyra.voice.tts.request","lyra.voice.tts.request.>","_INBOX.voice-tts.>","_inbox.voice-tts.>"] }
+        subscribe: { allow: ["lyra.voice.tts.request","lyra.voice.tts.request.>","_INBOX.voice-tts.>","_inbox.voice-tts.>"] }  # lowercase variant for nats.py <2.2 compatibility
         allow_responses: true
       }
     }
@@ -68,7 +68,7 @@ authorization {
       # voice-stt
       permissions: {
         publish:   { allow: ["lyra.voice.stt.heartbeat","lyra.system.ready"] }
-        subscribe: { allow: ["lyra.voice.stt.request","lyra.voice.stt.request.>","_INBOX.voice-stt.>","_inbox.voice-stt.>"] }
+        subscribe: { allow: ["lyra.voice.stt.request","lyra.voice.stt.request.>","_INBOX.voice-stt.>","_inbox.voice-stt.>"] }  # lowercase variant for nats.py <2.2 compatibility
         allow_responses: true
       }
     }
@@ -78,7 +78,7 @@ authorization {
       permissions: {
         publish:   { allow: ["lyra.llm.health.*"] }
         subscribe: { allow: ["lyra.llm.request"] }
-        allow_responses: true
+        allow_responses: true  # intentional: uses allow_responses only; nats_connect() must NOT use identity_name=
       }
     }
     {
@@ -86,7 +86,7 @@ authorization {
       # image-worker
       permissions: {
         publish:   { allow: ["lyra.image.heartbeat"] }
-        subscribe: { allow: ["lyra.image.generate.request","_INBOX.image-worker.>","_inbox.image-worker.>"] }
+        subscribe: { allow: ["lyra.image.generate.request","_INBOX.image-worker.>","_inbox.image-worker.>"] }  # lowercase variant for nats.py <2.2 compatibility
         allow_responses: true
       }
     }
@@ -96,7 +96,7 @@ authorization {
       permissions: {
         publish:   { allow: ["lyra.monitor.>"] }
         subscribe: { allow: ["lyra.monitor.>"] }
-        allow_responses: true
+        allow_responses: true  # intentional: uses allow_responses only; nats_connect() must NOT use identity_name=
       }
     }
     {
@@ -104,7 +104,7 @@ authorization {
       # clipool-worker
       permissions: {
         publish:   { allow: ["lyra.clipool.heartbeat","lyra.system.ready"] }
-        subscribe: { allow: ["lyra.clipool.cmd","lyra.clipool.control","_INBOX.clipool-worker.>","_inbox.clipool-worker.>"] }
+        subscribe: { allow: ["lyra.clipool.cmd","lyra.clipool.control","_INBOX.clipool-worker.>","_inbox.clipool-worker.>"] }  # lowercase variant for nats.py <2.2 compatibility
         allow_responses: true
       }
     }
