+++ scripts/check-acl-authconf-drift.sh
@@ -24,11 +24,11 @@ for f in "$AUTH_CONF" "$ACL_MATRIX" "$GEN_NKEYS"; do
 done
 
 # Create temp directory for normalized outputs
-TMPDIR_LOCAL=$(mktemp -d)
-trap 'rm -rf "$TMPDIR_LOCAL"' EXIT
+NATS_TMPDIR=$(mktemp -d)
+trap 'rm -rf "$NATS_TMPDIR"' EXIT
 
-COMMITTED_NORM="${TMPDIR_LOCAL}/committed_norm.txt"
-GENERATED_NORM="${TMPDIR_LOCAL}/generated_norm.txt"
+COMMITTED_NORM="${NATS_TMPDIR}/committed_norm.txt"
+GENERATED_NORM="${NATS_TMPDIR}/generated_norm.txt"
 
 # ---------------------------------------------------------------------------
 # normalize_authconf: strip comments, normalize nkey values, normalize whitespace
@@ -37,10 +37,12 @@ GENERATED_NORM="${TMPDIR_LOCAL}/generated_norm.txt"
 # ---------------------------------------------------------------------------
 normalize_authconf() {
   # 1. Strip comment lines (lines starting with #, ignoring leading whitespace)
-  # 2. Normalize nkey values: replace any nkey value with "NKEY_PLACEHOLDER"
-  # 3. Normalize whitespace: squeeze multiple spaces, trim trailing
+  # 2. Strip trailing inline comments (safe: NATS conf quoted strings don't contain #)
+  # 3. Normalize nkey values: replace any nkey value with "NKEY_PLACEHOLDER"
+  # 4. Normalize whitespace: squeeze multiple spaces, trim trailing
   sed -E \
     -e '/^[[:space:]]*#/d' \
+    -e 's/[[:space:]]*#[^"]*$//' \
     -e 's/nkey: "[^"]+"/nkey: "NKEY_PLACEHOLDER"/g' \
     -e 's/[[:space:]]+/ /g' \
     -e 's/[[:space:]]+$//' \
@@ -48,17 +50,19 @@ normalize_authconf() {
 }
 
 # Generate expected auth.conf from acl-matrix.json
-bash "$GEN_NKEYS" --template-only > "${TMPDIR_LOCAL}/generated_auth.conf"
+bash "$GEN_NKEYS" --template-only > "${NATS_TMPDIR}/generated_auth.conf"
 
 # Normalize both files
 normalize_authconf < "$AUTH_CONF" > "$COMMITTED_NORM"
-normalize_authconf < "${TMPDIR_LOCAL}/generated_auth.conf" > "$GENERATED_NORM"
+normalize_authconf < "${NATS_TMPDIR}/generated_auth.conf" > "$GENERATED_NORM"
 
-# Diff — exit 0 on match, 1 on drift
-if diff -u "$COMMITTED_NORM" "$GENERATED_NORM"; then
+# Diff — explicit capture makes intent clear under set -euo pipefail
+diff_out=$(diff -u "$COMMITTED_NORM" "$GENERATED_NORM" || true)
+if [ -z "$diff_out" ]; then
   echo "✓ auth.conf is in sync with acl-matrix.json"
   exit 0
 else
+  echo "$diff_out"
   echo "::error::auth.conf is out of sync with acl-matrix.json — run 'bash deploy/nats/gen-nkeys.sh --template-only > deploy/nats/auth.conf' and commit the result" >&2
   exit 1
 fi
