+++ src/lyra/adapters/discord/discord_normalize.py
@@ -30,8 +30,7 @@ def normalize(  # noqa: PLR0913 — all kwargs are platform-specific routing con
     thread_id: int | None = None,
     channel_id: int | None = None,
     trust_level: TrustLevel = TrustLevel.TRUSTED,
-    # REQUIRED: always pass is_admin=identity.is_admin — do not rely on default
-    is_admin: bool = False,
+    is_admin: bool = False,  # REQUIRED: always pass is_admin=identity.is_admin
 ) -> InboundMessage:
     """Convert a discord.py Message (or SimpleNamespace) to InboundMessage."""
     is_mention = adapter._bot_user is not None and adapter._bot_user in raw.mentions
