+++ src/lyra/core/cli/cli_protocol.py
@@ -1,149 +1,29 @@
 """NDJSON protocol layer for the Claude CLI subprocess.
 
-Pure I/O protocol concerns: payload serialisation, event parsing, timeout
-handling, command construction, and the result state machine.  No pool
-state is held here — all pool-specific context is passed as explicit
-arguments.
+Public re-export facade.  Shared types live in cli_protocol_types; the
+protocol implementations live in cli_non_streaming and cli_streaming.
+Import from this module for the full public surface.
 """
 
 from __future__ import annotations
 
-import asyncio
-import logging
-import os
-import re
-import tempfile
-from dataclasses import dataclass
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from .cli_non_streaming import read_until_result, send_and_read
-    from .cli_streaming import StreamingIterator, send_and_read_stream
-
-from ..agent.agent_config import ModelConfig
-
-log = logging.getLogger(__name__)
-
-
-# ---------------------------------------------------------------------------
-# Subprocess helpers
-# ---------------------------------------------------------------------------
-
-
-async def _read_stderr_snippet(
-    proc: asyncio.subprocess.Process, limit: int = 512
-) -> str:
-    """Read up to *limit* bytes from proc.stderr without blocking.
-
-    Returns a stripped string, or "" if nothing is available.
-    """
-    if proc.stderr is None:
-        return ""
-    try:
-        raw = await asyncio.wait_for(proc.stderr.read(limit), timeout=0.5)
-        return raw.decode(errors="replace").strip()
-    except (asyncio.TimeoutError, Exception):
-        return ""
-
-
-def build_cmd(
-    model_config: ModelConfig,
-    session_id: str | None = None,
-    system_prompt: str = "",
-) -> tuple[list[str], str | None]:
-    """Build the ``claude`` CLI command list from *model_config*.
-
-    Returns ``(cmd, prompt_file)`` where *prompt_file* is a temp path
-    that the caller must clean up, or ``None`` when no system prompt
-    was written.
-    """
-    cmd = [
-        "claude",
-        "--input-format",
-        "stream-json",
-        "--output-format",
-        "stream-json",
-        "--verbose",
-        "--model",
-        model_config.model,
-        # max_turns=None → unlimited; 0 is DB sentinel for None.
-        *(
-            ["--max-turns", str(model_config.max_turns)]
-            if model_config.max_turns
-            else []
-        ),
-    ]
-    if model_config.streaming:
-        cmd.append("--include-partial-messages")
-    if model_config.skip_permissions:
-        log.warning(
-            "SECURITY: --dangerously-skip-permissions enabled for CLI subprocess"
-        )
-        cmd.append("--dangerously-skip-permissions")
-    if model_config.tools:
-        cmd.extend(["--allowedTools", ",".join(model_config.tools)])
-    prompt_file: str | None = None
-    if system_prompt:
-        fd, prompt_file = tempfile.mkstemp(suffix=".txt", prefix="lyra-prompt-")
-        try:
-            os.write(fd, system_prompt.encode("utf-8"))
-        finally:
-            os.close(fd)
-        os.chmod(prompt_file, 0o600)
-        cmd.extend(["--system-prompt-file", prompt_file])
-    if session_id:
-        cmd.extend(["--resume", session_id])
-    return cmd, prompt_file
-
-
-# Validate Claude session IDs (strict UUID format).
-SESSION_ID_RE = re.compile(
-    r"^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
+from .cli_non_streaming import read_until_result, send_and_read
+from .cli_protocol_types import (
+    _SESSION_ID_RE,
+    SESSION_ID_RE,
+    CliProtocolOptions,
+    CliResult,
+    _read_stderr_snippet,
+    build_cmd,
 )
-# Private alias kept for backward compat (cli_pool.py imports this name).
-_SESSION_ID_RE = SESSION_ID_RE
-
-
-@dataclass
-class CliResult:
-    """Result from a CliPool.send() call.
-
-    Exactly one of ``result`` or ``error`` will be set (not both).
-    ``warning`` is set when the response was truncated (e.g. max_turns reached).
-    ``session_id`` is always present on success.
-    """
-
-    result: str = ""
-    session_id: str = ""
-    error: str = ""
-    warning: str = ""
-
-    @property
-    def ok(self) -> bool:
-        return not self.error
-
-
-@dataclass
-class CliProtocolOptions:
-    """Wire-level timing and retry options for the NDJSON protocol.
-
-    Carries the three ``[cli_pool]`` knobs from ``config.toml`` into the
-    protocol layer.  Defaults mirror pre-#369 hardcoded values.
-    """
-
-    stdin_drain_timeout: float = 10.0
-    max_idle_retries: int = 3
-    intermediate_timeout: float = 5.0
-
-
-# Re-export from submodules for backward compatibility (after class definitions)
-from .cli_non_streaming import read_until_result, send_and_read  # noqa: E402, I001 — circular import: cli_non_streaming/cli_streaming import CliProtocolOptions from this module
-from .cli_streaming import StreamingIterator, send_and_read_stream  # noqa: E402, I001 — circular import: cli_non_streaming/cli_streaming import CliProtocolOptions from this module
+from .cli_streaming import StreamingIterator, send_and_read_stream
 
 __all__ = [
     "CliProtocolOptions",
     "CliResult",
     "SESSION_ID_RE",
+    "_SESSION_ID_RE",
+    "_read_stderr_snippet",
     "StreamingIterator",
     "build_cmd",
     "read_until_result",
