+++ src/lyra/core/cli/cli_streaming.py
@@ -12,7 +12,7 @@ import logging
 from collections.abc import Awaitable, Callable
 
 from ..messaging.events import LlmEvent
-from .cli_protocol import CliProtocolOptions, _read_stderr_snippet
+from .cli_protocol_types import CliProtocolOptions, _read_stderr_snippet
 from .cli_streaming_parser import CliStreamingParser
 
 log = logging.getLogger(__name__)
